+++ xoto3/dynamodb/write_versioned/__init__.py
@@ -1,11 +1,24 @@
-"""API for transact_write_items"""
-from .errors import TableSchemaUnknownError, TransactionAttemptsOverrun  # noqa
+"""API for transact_write_items.
+
+Anything you import from any underlying module directly (i.e. not from
+here) is not guaranteed to remain, so don't do that. Import this
+module and use only what it exposes.
+"""
+from .api2 import ItemTable, TypedTable, create_or_update, update_existing, update_if_exists  # noqa
+from .errors import (  # noqa
+    ItemUndefinedException,
+    TableSchemaUnknownError,
+    TransactionAttemptsOverrun,
+)
 from .modify import delete, put  # noqa
 from .read import get, require  # noqa
+from .retry import timed_retry  # noqa
 from .run import versioned_transact_write_items  # noqa
+from .specify import define_table, presume  # noqa
 from .types import (  # noqa
     ItemKeysByTableName,
     ItemsByTableName,
+    TransactionBuilder,
     TransactWriteItems,
     VersionedTransaction,
 )