+++ CHANGES.md
@@ -1,3 +1,9 @@
+## 1.12
+
+An enhanced API supporting the use of
+`dynamodb.write_versioned.versioned_transact_write_items`, which is a
+great way to write business logic against DynamoDB.
+
 ### 1.11.1
 
 Change all Resources and Clients to use a thread-local Session to